backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
openai==1.54.0
pydantic==2.9.2
langgraph>=0.2.0
langchain-core>=0.3.0
typing-extensions>=4.9.0
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
